safari-600.7-branch/LayoutTests/http/tests/security/no-javascript-location-percent-escaped-expected.txt

PASS: This is the content that appears in place of a Location redirect to a javascript: URL.
